+++ Assets/CoreAiUnity/Tests/PlayMode/FastNoLlm/LuaBindingsIntegrationPlayModeTests.cs
@@ -1,3 +1,4 @@
+#if !UNITY_WEBGL
 using System.Collections.Generic;
 using CoreAI.Ai;
 using CoreAI.Infrastructure.Logging;
@@ -76,3 +77,4 @@ namespace CoreAI.Tests.PlayMode
         }
     }
 }
+#endif