+++ Assets/CoreAiUnity/Tests/PlayMode/FastNoLlm/LuaCoroutineRunnerPlayModeTests.cs
@@ -1,3 +1,4 @@
+#if !UNITY_WEBGL
 using System.Collections;
 using CoreAI.Infrastructure.Lua;
 using CoreAI.Sandbox;
@@ -117,3 +118,4 @@ namespace CoreAI.Tests.PlayMode
         }
     }
 }
+#endif