+++ Assets/CoreAiUnity/Tests/PlayMode/FastNoLlm/LuaFormulaRuntimeIntegrationPlayModeTests.cs
@@ -10,6 +10,7 @@ using UnityEngine;
 using UnityEngine.TestTools;
 using static CoreAI.Messaging.AiGameCommandTypeIds;
 
+#if !UNITY_WEBGL
 namespace CoreAI.Tests.PlayMode
 {
     /// <summary>
@@ -368,3 +369,4 @@ namespace CoreAI.Tests.PlayMode
         }
     }
 }
+#endif